LRQA-33092 Refactor Failure message classes into a failure-message-generator package
