+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/SubrepositoryGitHubMessageUtil.java
@@ -14,6 +14,8 @@
 
 package com.liferay.jenkins.results.parser;
 
+import com.liferay.jenkins.results.parser.failure.message.generator.FailureMessageUtil;
+
 import java.io.File;
 
 import org.apache.tools.ant.Project;
