+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/SubrepositoryTaskNoReport.java
@@ -14,6 +14,8 @@
 
 package com.liferay.jenkins.results.parser;
 
+import com.liferay.jenkins.results.parser.failure.message.generator.GenericFailureMessageGenerator;
+
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
