+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/UnstableMessageUtil.java
@@ -14,6 +14,8 @@
 
 package com.liferay.jenkins.results.parser;
 
+import com.liferay.jenkins.results.parser.failure.message.generator.GenericFailureMessageGenerator;
+
 import java.util.ArrayList;
 import java.util.List;
