+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/failure/message/generator/GenericFailureMessageGenerator.java
@@ -12,7 +12,9 @@
  * details.
  */
 
-package com.liferay.jenkins.results.parser;
+package com.liferay.jenkins.results.parser.failure.message.generator;
+
+import com.liferay.jenkins.results.parser.Build;
 
 import java.util.Hashtable;
